# use `xs[xs.length] = x` rather than `xs.push(x)`