--- a/CONTRIBUTING.md
+++ b/CONTRIBUTING.md
@@ -20,7 +20,7 @@
     `scripts/build --complete` to a temporary file, then rename the file.
 
 5.  Run `make test lint`, (or `grunt test`) and address any errors. Preferably,
-    fix commits in place using `git rebase` or `git commit --amend` to make the 
+    fix commits in place using `git rebase` or `git commit --amend` to make the
     changes easier to review and to keep the history tidy.
 
 6.  Push to your fork:
@@ -28,3 +28,11 @@
         $ git push origin <branch>
 
 7.  Open a pull request.
+
+### Conventions
+
+  - Do not use `Array.prototype.push`. Use `xs[idx] = x` or `xs[xs.length] = x`
+    rather than `xs.push(x)`. See [#890][1] for links to related pull requests.
+
+
+[1]: https://github.com/ramda/ramda/issues/890