--- a/src/differenceWith.js
+++ b/src/differenceWith.js
@@ -31,7 +31,7 @@ module.exports = _curry3(function differenceWith(pred, first, second) {
     var containsPred = containsWith(pred);
     while (++idx < firstLen) {
         if (!containsPred(first[idx], second) && !containsPred(first[idx], out)) {
-            out.push(first[idx]);
+            out[idx] = first[idx];
         }
     }
     return out;