--- a/src/internal/_filter.js
+++ b/src/internal/_filter.js
@@ -2,7 +2,7 @@ module.exports = function _filter(fn, list) {
     var idx = -1, len = list.length, result = [];
     while (++idx < len) {
         if (fn(list[idx])) {
-            result.push(list[idx]);
+            result[result.length] = list[idx];
         }
     }
     return result;