--- a/src/internal/_filterIndexed.js
+++ b/src/internal/_filterIndexed.js
@@ -2,7 +2,7 @@ module.exports = function _filterIndexed(fn, list) {
     var idx = -1, len = list.length, result = [];
     while (++idx < len) {
         if (fn(list[idx], idx, list)) {
-            result.push(list[idx]);
+            result[result.length] = list[idx];
         }
     }
     return result;