--- a/src/internal/_map.js
+++ b/src/internal/_map.js
@@ -1,7 +1,7 @@
 module.exports = function _map(fn, list) {
     var idx = -1, len = list.length, result = [];
     while (++idx < len) {
-        result.push(fn(list[idx]));
+        result[idx] = fn(list[idx]);
     }
     return result;
 };