--- a/src/intersectionWith.js
+++ b/src/intersectionWith.js
@@ -44,7 +44,7 @@ module.exports = _curry3(function intersectionWith(pred, list1, list2) {
     var results = [], idx = -1;
     while (++idx < list1.length) {
         if (_containsWith(pred, list1[idx], list2)) {
-            results.push(list1[idx]);
+            results[results.length] = list1[idx];
         }
     }
     return uniqWith(pred, results);