--- a/src/keysIn.js
+++ b/src/keysIn.js
@@ -23,7 +23,7 @@ var _curry1 = require('./internal/_curry1');
 module.exports = _curry1(function keysIn(obj) {
     var prop, ks = [];
     for (prop in obj) {
-        ks.push(prop);
+        ks[ks.length] = prop;
     }
     return ks;
 });