--- a/src/mapAccum.js
+++ b/src/mapAccum.js
@@ -29,7 +29,7 @@ module.exports = _curry3(function mapAccum(fn, acc, list) {
     var idx = -1, len = list.length, result = [], tuple = [acc];
     while (++idx < len) {
         tuple = fn(tuple[0], list[idx]);
-        result.push(tuple[1]);
+        result[idx] = tuple[1];
     }
     return [tuple[0], result];
 });