--- a/src/mapIndexed.js
+++ b/src/mapIndexed.js
@@ -30,7 +30,7 @@ var _curry2 = require('./internal/_curry2');
 module.exports = _curry2(function mapIndexed(fn, list) {
     var idx = -1, len = list.length, result = [];
     while (++idx < len) {
-        result.push(fn(list[idx], idx, list));
+        result[idx] = fn(list[idx], idx, list);
     }
     return result;
 });