--- a/src/partition.js
+++ b/src/partition.js
@@ -21,7 +21,8 @@ var _reduce = require('./internal/_reduce');
  */
 module.exports = _curry2(function partition(pred, list) {
     return _reduce(function(acc, elt) {
-        acc[pred(elt) ? 0 : 1].push(elt);
+        var xs = acc[pred(elt) ? 0 : 1];
+        xs[xs.length] = elt;
         return acc;
     }, [[], []], list);
 });