--- a/src/unfold.js
+++ b/src/unfold.js
@@ -27,7 +27,7 @@ module.exports = _curry2(function unfold(fn, seed) {
     var pair = fn(seed);
     var result = [];
     while (pair && pair.length) {
-        result.push(pair[0]);
+        result[result.length] = pair[0];
         pair = fn(pair[1]);
     }
     return result;