--- a/src/useWith.js
+++ b/src/useWith.js
@@ -73,7 +73,7 @@ module.exports = curry(function useWith(fn /*, transformers */) {
     return curry(arity(tlen, function() {
         var args = [], idx = -1;
         while (++idx < tlen) {
-            args.push(transformers[idx](arguments[idx]));
+            args[idx] = transformers[idx](arguments[idx]);
         }
         return fn.apply(this, args.concat(_slice(arguments, tlen)));
     }));