--- a/src/valuesIn.js
+++ b/src/valuesIn.js
@@ -23,7 +23,7 @@ var _curry1 = require('./internal/_curry1');
 module.exports = _curry1(function valuesIn(obj) {
     var prop, vs = [];
     for (prop in obj) {
-        vs.push(obj[prop]);
+        vs[vs.length] = obj[prop];
     }
     return vs;
 });