--- a/src/zip.js
+++ b/src/zip.js
@@ -23,7 +23,7 @@ module.exports = _curry2(function zip(a, b) {
     var idx = -1;
     var len = Math.min(a.length, b.length);
     while (++idx < len) {
-        rv.push([a[idx], b[idx]]);
+        rv[idx] = [a[idx], b[idx]];
     }
     return rv;
 });