Closes #890

This pull request:

  - reverts #887;
  - replaces the handful of uses of `Array.prototype.push` which predated #887; and
  - adds a note to a new "Conventions" section of __CONTRIBUTING.md__.

Before:

```
~/github.com/ramda/ramda (master)
$ grunt bench
Running "benchmark:all" (benchmark) task

Running suite concat [lib/bench/concat.bench.js]...
>> _(s1).concat(s2) x 1,368,969 ops/sec ±3.22% (95 runs sampled)
>> concat(s1, s2) x 403,106 ops/sec ±7.09% (90 runs sampled)
>> concat(s1)(s2) x 426,277 ops/sec ±1.84% (94 runs sampled)
>> concatS1(s2) x 453,114 ops/sec ±0.48% (97 runs sampled)
Fastest test is _(s1).concat(s2) at 3.0x faster than concatS1(s2)

Running suite curry [lib/bench/curry.bench.js]...
>> _x4(100) x 13,037,423 ops/sec ±1.59% (95 runs sampled)
>> mult4(100) x 1,241,952 ops/sec ±1.48% (94 runs sampled)
>> manual(100) x 61,685,048 ops/sec ±1.60% (91 runs sampled)
Fastest test is manual(100) at 4.7x faster than _x4(100)

Running suite forEach [lib/bench/each.bench.js]...
>> _.each(nums, x2) x 415,982 ops/sec ±0.61% (99 runs sampled)
>> forEach(x2, nums) x 423,948 ops/sec ±0.71% (98 runs sampled)
>> forEach(x2)(nums) x 414,706 ops/sec ±0.57% (97 runs sampled)
Fastest test is forEach(x2, nums) at 1.02x faster than forEach(x2)(nums) and _.each(nums, x2)

Running suite filter where [lib/bench/filter-where.bench.js]...
>> _.filter(objs, {x: []}) x 115,878 ops/sec ±1.68% (98 runs sampled)
>> filter(where({x: isEmpty}), objs) x 77,985 ops/sec ±2.07% (94 runs sampled)
>> filter(where({x: isEmpty}))(objs) x 66,790 ops/sec ±6.04% (88 runs sampled)
>> filterEmptyX(objs) x 69,758 ops/sec ±2.92% (89 runs sampled)
>> _.filter(objs, {y: false}) x 920,353 ops/sec ±2.36% (91 runs sampled)
>> filter(where({y: false}), objs) x 69,271 ops/sec ±0.53% (98 runs sampled)
>> filter(where({y: false}))(objs) x 56,283 ops/sec ±3.37% (82 runs sampled)
>> filterFalseY(objs) x 81,667 ops/sec ±2.71% (95 runs sampled)
Fastest test is _.filter(objs, {y: false}) at 7.9x faster than _.filter(objs, {x: []})

Running suite filter [lib/bench/filter.bench.js]...
>> _.filter(nums, isEven) x 2,613,794 ops/sec ±0.47% (98 runs sampled)
>> filter(isEven, nums) x 564,926 ops/sec ±2.36% (85 runs sampled)
>> filter(isEven)(nums) x 457,530 ops/sec ±10.90% (77 runs sampled)
>> filterEven(nums) x 606,480 ops/sec ±2.32% (93 runs sampled)
Fastest test is _.filter(nums, isEven) at 4.3x faster than filterEven(nums)

Running suite find where [lib/bench/find-where.bench.js]...
>> _.find(objs, {x: []}) x 625,614 ops/sec ±1.79% (95 runs sampled)
>> find(where({x: isEmpty}), objs) x 202,359 ops/sec ±3.36% (86 runs sampled)
>> find(where({x: isEmpty}))(objs) x 246,246 ops/sec ±0.75% (91 runs sampled)
>> findEmptyX(objs) x 274,585 ops/sec ±0.55% (101 runs sampled)
>> _.find(objs, {y: false}) x 1,232,320 ops/sec ±2.66% (90 runs sampled)
>> find(where({y: false}), objs) x 227,554 ops/sec ±0.78% (92 runs sampled)
>> find(where({y: false}))(objs) x 202,823 ops/sec ±4.57% (87 runs sampled)
>> findFalseY(objs) x 195,202 ops/sec ±5.49% (76 runs sampled)
Fastest test is _.find(objs, {y: false}) at 1.97x faster than _.find(objs, {x: []})

Running suite find [lib/bench/find.bench.js]...
>> _.find(nums, isZero) x 5,952,412 ops/sec ±2.32% (88 runs sampled)
>> find(isZero, nums) x 9,722,953 ops/sec ±2.37% (89 runs sampled)
>> find(isZero)(nums) x 5,128,579 ops/sec ±7.35% (73 runs sampled)
>> findZero(nums) x 10,596,381 ops/sec ±3.06% (90 runs sampled)
Fastest test is findZero(nums) at 1.09x faster than find(isZero, nums)

Running suite findIndex where [lib/bench/findIndex-where.bench.js]...
>> _.findIndex(objs, {x: []}) x 549,893 ops/sec ±5.52% (83 runs sampled)
>> findIndex(where({x: isEmpty}), objs) x 229,476 ops/sec ±3.21% (89 runs sampled)
>> findIndex(where({x: isEmpty}))(objs) x 196,259 ops/sec ±4.86% (78 runs sampled)
>> findIndexEmptyX(objs) x 189,437 ops/sec ±12.95% (68 runs sampled)
>> _.findIndex(objs, {y: false}) x 1,145,725 ops/sec ±5.89% (84 runs sampled)
>> findIndex(where({y: false}), objs) x 196,104 ops/sec ±6.59% (83 runs sampled)
>> findIndex(where({y: false}))(objs) x 215,852 ops/sec ±4.13% (90 runs sampled)
>> findIndexFalseY(objs) x 243,758 ops/sec ±0.53% (97 runs sampled)
Fastest test is _.findIndex(objs, {y: false}) at 2.1x faster than _.findIndex(objs, {x: []})

Running suite findIndex [lib/bench/findIndex.bench.js]...
>> _.findIndex(nums, isZero) x 6,575,388 ops/sec ±0.46% (98 runs sampled)
>> findIndex(isZero, nums) x 11,027,587 ops/sec ±1.10% (97 runs sampled)
>> findIndex(isZero)(nums) x 6,063,651 ops/sec ±2.68% (92 runs sampled)
>> findIndexZero(nums) x 11,222,322 ops/sec ±2.21% (91 runs sampled)
Fastest test is findIndexZero(nums) at 1.02x faster than findIndex(isZero, nums)

Running suite indexOf [lib/bench/indexof.bench.js]...
>> _.indexOf x 19,635,066 ops/sec ±6.54% (74 runs sampled)
>> indexOf(sq, nums) x 14,489,096 ops/sec ±7.96% (77 runs sampled)
>> indexOf(sq)(nums) x 7,415,894 ops/sec ±3.14% (83 runs sampled)
>> idxOf23(nums) x 20,006,247 ops/sec ±1.38% (94 runs sampled)
Fastest test is idxOf23(nums) at 1.02x faster than _.indexOf

Running suite isSet [lib/bench/isset.bench.js]...
>> isSet(unsortedBag) x 17,000,501 ops/sec ±2.13% (91 runs sampled)
>> isSet(unsortedSet) x 5,283,432 ops/sec ±2.04% (95 runs sampled)
>> isSet(sortedBag) x 16,638,628 ops/sec ±1.83% (89 runs sampled)
>> isSet(sortedSet) x 5,271,099 ops/sec ±1.22% (95 runs sampled)
Fastest test is isSet(unsortedBag) at 1.02x faster than isSet(sortedBag)

Running suite map [lib/bench/map.bench.js]...
>> _.map x 4,084,769 ops/sec ±1.86% (93 runs sampled)
>> map(sq, nums) x 377,451 ops/sec ±2.18% (91 runs sampled)
>> map(sq)(nums) x 391,288 ops/sec ±1.17% (93 runs sampled)
>> mapSq(nums) x 402,405 ops/sec ±1.37% (95 runs sampled)
Fastest test is _.map at 10.2x faster than mapSq(nums)

Running suite maxBy [lib/bench/maxWith.bench.js]...
>> _.max x 117,517 ops/sec ±2.05% (91 runs sampled)
>> maxBy(computer, nums) x 262,155 ops/sec ±1.57% (96 runs sampled)
>> maxBy(computer)(vals) x 264,003 ops/sec ±0.81% (98 runs sampled)
>> maxVal(vals) x 274,450 ops/sec ±0.41% (101 runs sampled)
Fastest test is maxVal(vals) at 1.05x faster than maxBy(computer, nums) and maxBy(computer)(vals)

Running suite reduce [lib/bench/reduce.bench.js]...
>> _.reduce(nums, add, 0) x 3,767,838 ops/sec ±5.60% (88 runs sampled)
>> reduce(add, 0, nums) x 6,703,083 ops/sec ±1.90% (92 runs sampled)
>> reduce(add, 0)(nums) x 4,315,366 ops/sec ±4.64% (79 runs sampled)
>> reduceAdd(nums) x 7,777,166 ops/sec ±1.92% (93 runs sampled)
Fastest test is reduceAdd(nums) at 1.16x faster than reduce(add, 0, nums)
```

After:

```
11:27 ~/github.com/ramda/ramda (push)
$ grunt bench
Running "benchmark:all" (benchmark) task

Running suite concat [lib/bench/concat.bench.js]...
>> _(s1).concat(s2) x 1,386,591 ops/sec ±1.80% (94 runs sampled)
>> concat(s1, s2) x 3,448,162 ops/sec ±1.23% (94 runs sampled)
>> concat(s1)(s2) x 2,671,543 ops/sec ±2.38% (89 runs sampled)
>> concatS1(s2) x 3,107,912 ops/sec ±3.64% (85 runs sampled)
Fastest test is concat(s1, s2) at 1.11x faster than concatS1(s2)

Running suite curry [lib/bench/curry.bench.js]...
>> _x4(100) x 11,149,670 ops/sec ±4.63% (87 runs sampled)
>> mult4(100) x 1,800,595 ops/sec ±5.01% (85 runs sampled)
>> manual(100) x 52,777,090 ops/sec ±2.21% (84 runs sampled)
Fastest test is manual(100) at 4.7x faster than _x4(100)

Running suite forEach [lib/bench/each.bench.js]...
>> _.each(nums, x2) x 402,820 ops/sec ±0.80% (95 runs sampled)
>> forEach(x2, nums) x 350,622 ops/sec ±4.57% (84 runs sampled)
>> forEach(x2)(nums) x 383,634 ops/sec ±2.77% (92 runs sampled)
Fastest test is _.each(nums, x2) at 1.15x faster than forEach(x2, nums) and forEach(x2)(nums)

Running suite filter where [lib/bench/filter-where.bench.js]...
>> _.filter(objs, {x: []}) x 90,490 ops/sec ±6.53% (80 runs sampled)
>> filter(where({x: isEmpty}), objs) x 85,409 ops/sec ±4.21% (85 runs sampled)
>> filter(where({x: isEmpty}))(objs) x 86,835 ops/sec ±5.87% (94 runs sampled)
>> filterEmptyX(objs) x 79,729 ops/sec ±3.95% (78 runs sampled)
>> _.filter(objs, {y: false}) x 768,583 ops/sec ±9.11% (76 runs sampled)
>> filter(where({y: false}), objs) x 55,069 ops/sec ±15.28% (71 runs sampled)
>> filter(where({y: false}))(objs) x 68,521 ops/sec ±4.07% (85 runs sampled)
>> filterFalseY(objs) x 81,487 ops/sec ±2.22% (95 runs sampled)
Fastest test is _.filter(objs, {y: false}) at 8.5x faster than _.filter(objs, {x: []})

Running suite filter [lib/bench/filter.bench.js]...
>> _.filter(nums, isEven) x 2,640,032 ops/sec ±1.37% (96 runs sampled)
>> filter(isEven, nums) x 2,533,728 ops/sec ±0.71% (98 runs sampled)
>> filter(isEven)(nums) x 1,795,904 ops/sec ±3.16% (86 runs sampled)
>> filterEven(nums) x 2,299,163 ops/sec ±3.05% (87 runs sampled)
Fastest test is _.filter(nums, isEven) at 1.04x faster than filter(isEven, nums)

Running suite find where [lib/bench/find-where.bench.js]...
>> _.find(objs, {x: []}) x 619,718 ops/sec ±1.85% (89 runs sampled)
>> find(where({x: isEmpty}), objs) x 292,321 ops/sec ±0.95% (93 runs sampled)
>> find(where({x: isEmpty}))(objs) x 268,966 ops/sec ±1.88% (93 runs sampled)
>> findEmptyX(objs) x 324,765 ops/sec ±0.91% (95 runs sampled)
>> _.find(objs, {y: false}) x 1,312,225 ops/sec ±0.51% (98 runs sampled)
>> find(where({y: false}), objs) x 253,730 ops/sec ±1.16% (94 runs sampled)
>> find(where({y: false}))(objs) x 241,227 ops/sec ±0.90% (93 runs sampled)
>> findFalseY(objs) x 262,049 ops/sec ±0.83% (96 runs sampled)
Fastest test is _.find(objs, {y: false}) at 2.1x faster than _.find(objs, {x: []})

Running suite find [lib/bench/find.bench.js]...
>> _.find(nums, isZero) x 5,681,946 ops/sec ±1.76% (87 runs sampled)
>> find(isZero, nums) x 9,029,246 ops/sec ±3.81% (90 runs sampled)
>> find(isZero)(nums) x 6,025,048 ops/sec ±1.33% (91 runs sampled)
>> findZero(nums) x 11,267,172 ops/sec ±1.27% (94 runs sampled)
Fastest test is findZero(nums) at 1.25x faster than find(isZero, nums)

Running suite findIndex where [lib/bench/findIndex-where.bench.js]...
>> _.findIndex(objs, {x: []}) x 421,490 ops/sec ±9.78% (67 runs sampled)
>> findIndex(where({x: isEmpty}), objs) x 260,069 ops/sec ±9.11% (92 runs sampled)
>> findIndex(where({x: isEmpty}))(objs) x 274,760 ops/sec ±1.27% (94 runs sampled)
>> findIndexEmptyX(objs) x 310,609 ops/sec ±1.50% (92 runs sampled)
>> _.findIndex(objs, {y: false}) x 1,081,033 ops/sec ±6.49% (81 runs sampled)
>> findIndex(where({y: false}), objs) x 235,401 ops/sec ±6.55% (90 runs sampled)
>> findIndex(where({y: false}))(objs) x 206,008 ops/sec ±5.96% (83 runs sampled)
>> findIndexFalseY(objs) x 231,775 ops/sec ±4.05% (86 runs sampled)
Fastest test is _.findIndex(objs, {y: false}) at 2.6x faster than _.findIndex(objs, {x: []})

Running suite findIndex [lib/bench/findIndex.bench.js]...
>> _.findIndex(nums, isZero) x 6,454,087 ops/sec ±0.64% (96 runs sampled)
>> findIndex(isZero, nums) x 10,386,156 ops/sec ±1.29% (93 runs sampled)
>> findIndex(isZero)(nums) x 6,086,126 ops/sec ±2.25% (93 runs sampled)
>> findIndexZero(nums) x 11,608,603 ops/sec ±1.06% (94 runs sampled)
Fastest test is findIndexZero(nums) at 1.12x faster than findIndex(isZero, nums)

Running suite indexOf [lib/bench/indexof.bench.js]...
>> _.indexOf x 25,862,557 ops/sec ±0.95% (93 runs sampled)
>> indexOf(sq, nums) x 19,114,500 ops/sec ±0.98% (98 runs sampled)
>> indexOf(sq)(nums) x 7,476,330 ops/sec ±2.51% (76 runs sampled)
>> idxOf23(nums) x 21,108,074 ops/sec ±0.45% (96 runs sampled)
Fastest test is _.indexOf at 1.23x faster than idxOf23(nums)

Running suite isSet [lib/bench/isset.bench.js]...
>> isSet(unsortedBag) x 17,319,852 ops/sec ±1.17% (94 runs sampled)
>> isSet(unsortedSet) x 5,414,832 ops/sec ±0.97% (93 runs sampled)
commit 52f32cf8f2e8aee6a6ae3202290d8848822d92d4
>> isSet(sortedBag) x 18,292,157 ops/sec ±0.67% (98 runs sampled)
>> isSet(sortedSet) x 5,401,572 ops/sec ±1.10% (94 runs sampled)
Fastest test is isSet(sortedBag) at 1.06x faster than isSet(unsortedBag)

Running suite map [lib/bench/map.bench.js]...
>> _.map x 4,261,501 ops/sec ±1.00% (96 runs sampled)
>> map(sq, nums) x 2,439,993 ops/sec ±1.27% (97 runs sampled)
>> map(sq)(nums) x 1,841,576 ops/sec ±1.88% (92 runs sampled)
>> mapSq(nums) x 2,440,210 ops/sec ±1.19% (95 runs sampled)
commit 52f32cf8f2e8aee6a6ae3202290d8848822d92d4
Fastest test is _.map at 1.75x faster than map(sq, nums) and mapSq(nums)

Running suite maxBy [lib/bench/maxWith.bench.js]...
>> _.max x 145,845 ops/sec ±0.88% (97 runs sampled)
>> maxBy(computer, nums) x 267,226 ops/sec ±1.16% (99 runs sampled)
>> maxBy(computer)(vals) x 266,891 ops/sec ±1.14% (95 runs sampled)
>> maxVal(vals) x 280,475 ops/sec ±0.95% (98 runs sampled)
Fastest test is maxVal(vals) at 1.05x faster than maxBy(computer)(vals) and maxBy(computer, nums)

Running suite reduce [lib/bench/reduce.bench.js]...
>> _.reduce(nums, add, 0) x 4,372,079 ops/sec ±0.40% (100 runs sampled)
>> reduce(add, 0, nums) x 7,233,313 ops/sec ±0.65% (99 runs sampled)
>> reduce(add, 0)(nums) x 4,847,441 ops/sec ±0.65% (88 runs sampled)
>> reduceAdd(nums) x 8,287,915 ops/sec ±0.68% (96 runs sampled)
Fastest test is reduceAdd(nums) at 1.15x faster than reduce(add, 0, nums)
```

The fact that `map'` is capable of six times as many operations per second as `map` is surprising.